requirements.txt
paho-mqtt==1.6.1
aiohttp==3.9.1
pyyaml==6.0.1
tenacity==8.2.3
python-dotenv==1.0.0 